adhering to review comments
